requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
supabase==2.0.0
openai==1.12.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
email-validator
stripe==7.8.0>=2.0.0
